#1  -  changed data type of userId across PostController, UserController to integrate changes with data.sql

+++ src/main/java/com/rev_connect_api/controllers/UserController.java
@@ -42,7 +42,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserResponseDTO> getUserById(@PathVariable BigInteger id) {
+    public ResponseEntity<UserResponseDTO> getUserById(@PathVariable Long id) {
         User user = userService.findUserById(id);
         return ResponseEntity.ok(mapToDTO(user));
     }
